services/web/requirements.txt
Flask-SQLAlchemy==3.0.2
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
SQLAlchemy==1.4.46
flask-login==0.6.2
jsonschema==4.17.3
flask_cors==3.0.10
python-dateutil==2.8.2
